Automatically Subscribe to Topic when Posting
In #79, it is suggested that we subscribe to a topic when a user contributes i.e posts/authors the topic. I have subscribed the user to the topic whenever they post on it, also if they are already subscribed don't subscribe a second time which would cause multiple emails.

Added new unit test to verify that a user can only be subscribed to a topic once.
All test run and pass locally.