Fix Favorite & Subscription Topic pages
The Favorites (`Forums/Favorites/Topics`) and Subscription (`Forums/Subscription/Topics`) pages allow you to un-favorite and unsubscribe respectively. These actions havn't worked in a while from what I can see because once the buttons are clicked the topic ID which should be passed back is not getting registered with the correct ID in the controller. We need to map the HTML input tag to the parameter in the controller action.

When trying to unsubscribe from a topic it just unsubscribes from the topic with the ID of the default value of controller action i.e 0. When you try and un-favorite a topic on this page it actually throws an exception because you not a favorite of topic 0.