Show user not found in PM with deleted user update tests #321
This took me hours because of ADHD and an inability to focus. So on the off chance that you happen to read this, as a consumer of open source software, please thank the contributors who make the stuff that you use the most. It ain't easy, and it's free to you.